+++ src/main/java/com/stefanmilojevic/myRealEstate/controller/AdvertisementPictureController.java
@@ -7,7 +7,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
 import java.util.List;
 
@@ -23,12 +22,12 @@ public class AdvertisementPictureController {
     }
 
     @GetMapping("/getAllByAdvertisementId/{id}")
-    public ResponseEntity<List<AdvertisementPicture>> getAllByAdvertisementId(@PathVariable int id) {
+    public ResponseEntity<List<AdvertisementPicture>> getAllByAdvertisementId(@PathVariable Long id) {
         return ResponseEntity.ok(advertisementPictureService.getAllByAdvertisementId(id));
     }
 
     @PostMapping("/{advertisementId}")
-    public ResponseEntity<String> save(@RequestBody MultipartFile file, @PathVariable int advertisementId) throws IOException {
+    public ResponseEntity<String> save(@RequestBody MultipartFile file, @PathVariable Long advertisementId) throws IOException {
         return ResponseEntity.ok(advertisementPictureService.save(file, advertisementId));
     }
 
@@ -38,12 +37,12 @@ public class AdvertisementPictureController {
     }
 
     @PutMapping("/setPrimary/{advertisementId}/{id}")
-    public ResponseEntity<AdvertisementPicture> setPrimary(@PathVariable int advertisementId, @PathVariable int id) {
+    public ResponseEntity<AdvertisementPicture> setPrimary(@PathVariable Long advertisementId, @PathVariable int id) {
         return ResponseEntity.ok(advertisementPictureService.setPrimaryById(advertisementId, id));
     }
 
     @GetMapping("/getPrimaryByAdvertisementId/{advertisementId}")
-    public ResponseEntity<AdvertisementPicture> getPrimaryByAdvertisementId(@PathVariable int advertisementId) {
+    public ResponseEntity<AdvertisementPicture> getPrimaryByAdvertisementId(@PathVariable Long advertisementId) {
         return ResponseEntity.ok(advertisementPictureService.getPrimaryByAdvertisementId(advertisementId));
     }
 }
