+++ src/main/java/com/stefanmilojevic/myRealEstate/repository/AdvertisementPictureRepository.java
@@ -1,6 +1,5 @@
 package com.stefanmilojevic.myRealEstate.repository;
 
-import com.stefanmilojevic.myRealEstate.model.Advertisement;
 import com.stefanmilojevic.myRealEstate.model.AdvertisementPicture;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -9,7 +8,7 @@ import java.util.List;
 
 @Repository
 public interface AdvertisementPictureRepository extends JpaRepository<AdvertisementPicture, Integer> {
-    List<AdvertisementPicture> findAllByAdvertisement_Id(int advertisementId);
+    List<AdvertisementPicture> findAllByAdvertisement_Id(Long advertisementId);
     AdvertisementPicture findById(int id);
-    AdvertisementPicture findByPrimaryPicIsTrueAndAdvertisement_Id(int advertisementId);
+    AdvertisementPicture findByPrimaryPicIsTrueAndAdvertisement_Id(Long advertisementId);
 }
